+++ lib/src/iso/bmff/Box.ts
@@ -1,0 +1,14 @@
+/**
+ * Box
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type Box<T = any> = T & {
+	type: string;
+	size: number;
+	largesize?: number;
+	usertype?: number[];
+	boxes?: Box[];
+}