+++ lib/src/iso/bmff/fields/DATA.ts
@@ -1,0 +1,8 @@
+/**
+ * The data field type
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export const DATA = 'data';