+++ lib/src/iso/bmff/fields/TEMPLATE.ts
@@ -1,0 +1,8 @@
+/**
+ * The template field type
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export const TEMPLATE = 'template';