+++ lib/src/iso/bmff/parsers/sthd.ts
@@ -1,0 +1,26 @@
+import type { FullBox } from '../FullBox';
+import type { IsoView } from '../IsoView';
+
+/**
+ * ISO/IEC 14496-12:2015 - 12.6.2 Subtitle media header Box
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type SubtitleMediaHeaderBox = FullBox;
+
+/**
+ * Parse a SubtitleMediaHeaderBox from an IsoView
+ *
+ * @param view - The IsoView to read data from
+ *
+ * @returns A parsed SubtitleMediaHeaderBox
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export function sthd(view: IsoView): SubtitleMediaHeaderBox {
+	return view.readFullBox();
+};