+++ lib/test/iso/bmff/util/load.ts
@@ -1,0 +1,5 @@
+import fs from 'node:fs';
+
+export function load(file: string): ArrayBuffer {
+	return new Uint8Array(fs.readFileSync(`./test/isobmff/fixtures/${file}`)).buffer;
+}